+++ src/Client/Core/OrchestrationMetadata.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Text;
 
@@ -78,6 +79,11 @@ public sealed class OrchestrationMetadata
     /// </summary>
     /// <value>The serialized custom status or <c>null</c>.</value>
     public string? SerializedCustomStatus { get; init; }
+
+    /// <summary>
+    /// Gets the tags associated with the orchestration instance.
+    /// </summary>
+    public IReadOnlyDictionary<string, string> Tags { get; init; } = ImmutableDictionary.Create<string, string>();
 
     /// <summary>
     /// Gets the failure details, if any, for the orchestration instance.