+++ src/Worker/Core/Shims/TaskOrchestrationShim.cs
@@ -18,6 +18,9 @@ partial class TaskOrchestrationShim : TaskOrchestration
 {
     readonly ITaskOrchestrator implementation;
     readonly OrchestrationInvocationContext invocationContext;
+    readonly ILogger logger;
+    readonly IReadOnlyDictionary<string, object?> properties;
+
     TaskOrchestrationContextWrapper? wrapperContext;
 
     /// <summary>
@@ -28,14 +31,29 @@ partial class TaskOrchestrationShim : TaskOrchestration
     public TaskOrchestrationShim(
         OrchestrationInvocationContext invocationContext,
         ITaskOrchestrator implementation)
+        : this(invocationContext, implementation, new Dictionary<string, object?>())
+    {
+    }
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="TaskOrchestrationShim"/> class.
+    /// </summary>
+    /// <param name="invocationContext">The invocation context for this orchestration.</param>
+    /// <param name="implementation">The orchestration's implementation.</param>
+    /// <param name="properties">Configuration for the orchestration.</param>
+    public TaskOrchestrationShim(
+        OrchestrationInvocationContext invocationContext,
+        ITaskOrchestrator implementation,
+        IReadOnlyDictionary<string, object?> properties)
     {
         this.invocationContext = Check.NotNull(invocationContext);
         this.implementation = Check.NotNull(implementation);
+        this.properties = Check.NotNull(properties);
+
+        this.logger = Logs.CreateWorkerLogger(this.invocationContext.LoggerFactory, "Orchestrations");
     }
 
     DataConverter DataConverter => this.invocationContext.Options.DataConverter;
-
-    ILoggerFactory LoggerFactory => this.invocationContext.LoggerFactory;
 
     /// <inheritdoc/>
     public override async Task<string?> Execute(OrchestrationContext innerContext, string rawInput)
@@ -46,17 +64,33 @@ partial class TaskOrchestrationShim : TaskOrchestration
         innerContext.ErrorDataConverter = converterShim;
 
         object? input = this.DataConverter.Deserialize(rawInput, this.implementation.InputType);
-        this.wrapperContext = new(innerContext, this.invocationContext, input);
+        this.wrapperContext = new(innerContext, this.invocationContext, input, this.properties);
+
+        string instanceId = innerContext.OrchestrationInstance.InstanceId;
+        if (!innerContext.IsReplaying)
+        {
+            this.logger.OrchestrationStarted(instanceId, this.invocationContext.Name);
+        }
 
         try
         {
             object? output = await this.implementation.RunAsync(this.wrapperContext, input);
+
+            if (!innerContext.IsReplaying)
+            {
+                this.logger.OrchestrationCompleted(instanceId, this.invocationContext.Name);
+            }
 
             // Return the output (if any) as a serialized string.
             return this.DataConverter.Serialize(output);
         }
         catch (TaskFailedException e)
         {
+            if (!innerContext.IsReplaying)
+            {
+                this.logger.OrchestrationFailed(e, instanceId, this.invocationContext.Name);
+            }
+
             // Convert back to something the Durable Task Framework natively understands so that
             // failure details are correctly propagated.
             throw new CoreTaskFailedException(e.Message, e.InnerException)